fix: traj array int

+++ agents/erode.py
@@ -87,7 +87,7 @@ class Agent(Base):
 
         # initialise trajectory and pi_action arrays
         trajs = np.zeros(
-            shape=(self.cfg.particles, self.cfg.popsize * (1 - self.cfg.mix_coeff), self.cfg.horizon,
+            shape=(self.cfg.particles, int(self.cfg.popsize * (1 - self.cfg.mix_coeff)), self.cfg.horizon,
                    self.obs_dim + self.cfg.time_dim))
         if pi:
             pi_acts = np.zeros(shape=(self.cfg.particles, self.cfg.popsize * self.cfg.mix_coeff, self.cfg.horizon,
